Fix behaviour of IO read instructions
Previously, spaceman was just reading the value from stdin and putting the
converted result onto the stack. But the spec says, it should store it in its
heap memory section: Take the desired heap address from the stack and store the
read result at that position. That's what this commit implements.

This makes spaceman run quite a number of examples from the web successfully.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>