Implement Slide and Copy stack operations
These were not present in the whitespace spec on the net, that I found. But
other interpreters seem to implement it. I stumbled across these when parsing a
brainfuck interpreter and a sudoku solver written in whitespace didn't work.

With these changes, the sudoku solver works. Didn't try the brainfuck interpre-
ter. It parses, though.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>