Add simple hspec based test-suite for spaceman
Doesn't look like cabal supports programs and libraries in the same source di-
rectory (or — maybe more likely — I don't know how), so SpaceMan.hs is not in
‘prg/’, which is just as well.

Rearranged cabal file to create a library that gets used by the application as
well as the new test-suite. The suite just checks a couple of simple things in
the Alphabet module, to see if things fall into place correctly.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>