Add tests for label-transformation internals, fix bug in process
I noticed generated labels to be prefixed by a NUL character. I thought this
might be another convention to introduce human-readable labels. But then I saw
this to be the case with code that SpaceMan.Generate produces. And it doesn't
do that on its own. So this was a bug.

The issue is with the initial accumulator to ‘split8’ being [[]] which is [""]
when it comes to [String] lists. And eightPackToInt "" is NUL. Mystery solved,
it was the janitor.

These tests already get a little repetitive. I should be using quickcheck for
some of them.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>