Add Label variant that boils down to an Integer
The idea here is, to add another transformation before loading a program into
the virtual machine. What's currently is being done, when we are encountering a
label in a program is to lookup that label in the label table and use the ad-
dress that reveals. It would be much more beneficial to replace all those look-
ups with the addresses they would translate to — since all those lookups are
static — before executing the program.

This puts the required infrastructure in place.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>